Sales data info during Diwali period is gathered.
Analysis on various factors that affect the sales is been done. We can clearly see the graphs and visual representation of the data. This analysis gives a clear idea about the categorical emphasis and data driven decisions to be taken on specific gender, occupation, product category and more factors to improve the sales in future.